ml-micro/requirements.txt
pyserial==3.5
pynput==1.7.6